requirements.txt
pytest>=7.0.0
selenium>=4.0.0
django>=4.2.0
psycopg2-binary>=2.9.0
requests>=2.28.0
beautifulsoup4>=4.11.0
pandas>=1.5.0
numpy>=1.24.0
matplotlib>=3.6.0
seaborn>=0.11.0
websockets>=10.4
aiohttp>=3.8.0
pyyaml>=6.0
python-dotenv>=0.19.0
coverage>=6.5.0
pytest-cov>=4.0.0
pytest-html>=3.1.0
pytest-xdist>=3.0.0
locust>=2.14.0
paramiko>=3.0.0
